Move popup src into src/popup/ module
- getting too complicated for single script
