+++ src/popup/index.js
@@ -1,0 +1,2 @@
+import './search-input'
+import './archive-button'
